#77 Project is now modularized, and tests now work
