+++ project/cliApplication/src/main/java/cli/Is2projectApplication.java
@@ -1,4 +1,4 @@
-package is2;
+package cli;
 
 import crawler.Crawler;
 import crawler.LinioCrawler;
