+++ project/cliApplication/src/test/java/cli/Is2projectApplicationTests.java
@@ -1,4 +1,4 @@
-package is2;
+package cli;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
