+++ project/persistence/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:postgresql://localhost:5432/DATABASE
-spring.datasource.username=USERNAME
-spring.datasource.password=PASSWORD
+spring.datasource.url=jdbc:postgresql://localhost:5432/is2_t1
+spring.datasource.username=postgres
+spring.datasource.password=falvarez
 
 # Don't push the modified version of this file!
 # use:
