+++ project/persistence/src/test/resources/application.properties
@@ -1,0 +1,1 @@
+spring.jpa.database=HSQL
