--- project/src/main/java/META-INF/MANIFEST.MF
@@ -1,2 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: is2.Is2projectApplication
