--- project/src/test/resources/application.properties
@@ -1,1 +1,0 @@
-spring.jpa.database: HSQL
