+++ project/webScraper/src/test/java/crawler/LinioCrawlerTest.java
@@ -1,0 +1,20 @@
+package crawler;
+
+import static org.junit.Assert.*;
+
+import org.junit.Test;
+
+// TODO: this is just for test that tests works xd, do the proper tests
+public class LinioCrawlerTest {
+
+  @Test
+  public void getHtmlFromURL() throws Exception {
+    assertTrue(true);
+  }
+
+  @Test
+  public void buildCategories() throws Exception {
+    assertTrue(true);
+  }
+
+}
